fix(admin): reset-password --prompt-password; share the TTY prompt helper
Closes the last live help-equivalence gap: legacy `kweaver-admin user
reset-password` offers `--prompt-password` to type the new password
interactively, which openbkn lacked.

- Add `--prompt-password` to `admin user reset-password` (hidden input).
- Extract the hidden-capable `promptLine` from commands/auth.ts into
  utils/prompt.ts and reuse it in both. No behaviour change for auth.

With this, `BKN_EQUIV_LIVE=1` help equivalence is fully green (196/196).
